+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- function Systran.Translate.translate() for text translation
+- Systran.Client module for http requests
+- function fetch() for post request in Systran.Client module
+
 ## [0.0.1] - 2019-08-15
 ### Added
 - new project